requirements.txt
Scrapy==2.11.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
schedule==1.2.2
sqlalchemy==2.0.31
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
setuptools==70.3.0
asyncpg==0.29.0
requests==2.32.3
numpy==2.0.0
pandas==2.2.2
streamlit==1.36.0
deep_translator==1.11.4